# perf: make screen size and position calculations more efficient