… runtime/doc/news.txt
… src/nvim/ops.c
… src/nvim/plines.c
… src/nvim/plines.h
… test/benchmark/screenpos_spec.lua