NOTE: all referenced commits have been shifted by one, first commit now contains benchmarks.

### 1. Breakindent and strcmp()
`win_lbr_chartabsize()` recomputes `breakindent` of a line up to 2 times in order to determine the size of the character. [The function for computing breakindent](https://github.com/neovim/neovim/blob/0a598c13b1863ba1aff378027c4376e3ab7048ee/src/nvim/indent.c#L785) caches the last result, but as part of checking if the cached value could be used it strcmp's the given line to the cached (copied) line. Since `win_lbr_chartabsize()` is often called in a loop and none of the parameters that affect `breakindent` change, this results in strcmp being called for 2 identical lines (worst case for strcmp) for every character in the line starting from the first wrapped row (even though the value is almost never affects the final size of that character).

The first commit fixes the ussue of repeatedly comparing the lines by caching the indentation value in the `chartabsize_T` struct. I also decided to preserve the size of the structure by removing `cts_has_virt_text` and changing `cts_row` to `virt_row` since it is used only for virtual text and can be set to negative value if the line does not contain virtual text.

### 2. Unnecessary loop
`getvcol()` computes virtual column until the last position by iterating until the current position reaches the first byte of the character containing the last position. To compute the first byte of the character, it uses an extra loop at the start to find if the position is past the end of the line and sets it to the end of the line before computing the index of its first byte.

The second commit substitutes comparison with the first byte of the last position with checking if next position is greater than the last position, which allows removing the loop at the start.

### 3. Non-ascii characters are decoded multiple times.
In `getvcol()`, when the size of a Unicode character is computed, the character is decoded and validated (all bytes are read again) once in [`utf_ptr2cells()`](https://github.com/neovim/neovim/blob/0a598c13b1863ba1aff378027c4376e3ab7048ee/src/nvim/mbyte.c#L512). Then, when the pointer is advanced to the next character, its length is computed again (by reading all its bytes) in [`utfc_ptr2len()`](https://github.com/neovim/neovim/blob/0a598c13b1863ba1aff378027c4376e3ab7048ee/src/nvim/mbyte.c#L848), and then it is (possibly) decoded again in [`utf_composinglike()`](https://github.com/neovim/neovim/blob/0a598c13b1863ba1aff378027c4376e3ab7048ee/src/nvim/mbyte.c#L712). The next character is decoded once in `utf_composinglike()`, then its length is computed in `utfc_ptr2len()`, then it is (possibly) decoded a second time if it is a composing character, or it goes though all the steps the first character went. Typically a single Unicode character will be decoded twice and its length will be computed twice per one iteration.
Additionally, the function for decoding a character ([`utf_ptr2char()`](https://github.com/neovim/neovim/blob/0a598c13b1863ba1aff378027c4376e3ab7048ee/src/nvim/mbyte.c#L598)) is and'ing all bytes with masks like `0x3f` to remove the utf8 `10` and `11` bits, but since we know their values, we can subtract these bits instead of and'ing them with masks, which makes it possible to combine all the subtracted bits into one constant.

The issue of duplicate decoding and checking is fixed by adding 3 new functions: `utf_ptr2CharInfo()`, `ptr2CharInfo()` and `utfc_next()`, and modifying `getvcol()`. The first function decodes a character like `utf_ptr2char()`, but this new function returns more information about the decoded character (like its length and whether it is valid). The second function uses the first function to decode a multibyte character, but does a quick return if the character is single byte. The third function works like `utfc_ptr2len()`, but reuses previously decoded characters and also returns the last decoded character which is used on the next iteration of 'getvcol()` loop. The second and third functions are always inline because majority of the time they will be used on ascii characters and not inlining them would add overhead of a function call for the most common usecase.

I verified that `ptr2CharInfo()` and the new version of `utf_ptr2char()` return correct code points for all characters 1-4 bytes long (and some 5 and 6 byte characters) and return a negative value/first byte otherwise.

### 4. Optimizing other places that compute text size.
If wrapping is enabled, `screenpos()` will compute the screen size of all lines above the current line (in [`plines_win_nofold()`](https://github.com/neovim/neovim/blob/bbd5c6363c25e8fbbfb962f8f6c5ea1800d431ca/src/nvim/plines.c#L796)), but only `getvcol()` has the optimized loop for computing character sizes. This function and many other have to use either `win_lbr_chartabsize()` or `lbr_chartabsize()` to compute text size or position. Although both functions have a fast path and check for the same\* conditions as `getvcol()`, all of the checks have to be performed on each iteration. Additionally, Unicode characters are decoded multiple times, like in `getvcol()`.

The fourth commit splits the character size functions into 'regular' and 'fast' versions (`win_lbr_chartabsize()` -> `charsize_regular()`, `win_nolbr_chartabsize()` ~> `charsize_fast()`, `win_linetabsize_cts()` -> `linesize_regular()`, `linesize_fast()`). Checks for which version to use are computed when 'charsize' argument is initialized.
Other functions inside plines.c also have access to the `int_charsize_fast()` which is always inlined.
Both versions of 'charsize' functions now also take decoded characters as arguments instead of decoding them from the pointer, which allows to reuse decoded characters in all other places.

### Other changes

1. [Condition in `getvcol()`](https://github.com/neovim/neovim/blob/ec7453804421f738955a5f4460f44172318d42ff/src/nvim/plines.c#L517) originally checked for tab behavior and used `win_lbr_chartabsize()` if tabs were drawn as '^I'. But `win_lbr_chartabsize()` didn't explicitly check for that and called either `win_chartabsize()` or `win_nolbr_chartabsize()`, both of which checked the same tab behavior as `getvcol()`, but in [`== TAB` condition](https://github.com/neovim/neovim/blob/ec7453804421f738955a5f4460f44172318d42ff/src/nvim/plines.c#L46). My version makes `getvcol()` use fast path regardless of whether tabs have variable width or are displayed as '^I'.
2. [Docs](https://github.com/neovim/neovim/blob/ec7453804421f738955a5f4460f44172318d42ff/runtime/doc/mbyte.txt#L664) say that 'overlong sequences are not handled specially and displayed like a valid character'. I don't know how relevant this is since many places that I modified didn't originally handle overlong ascii characters as ascii (e.g. [`== TAB` in getvcol](https://github.com/neovim/neovim/blob/ec7453804421f738955a5f4460f44172318d42ff/src/nvim/plines.c#L534) and other places, [BYTE2LEN](https://github.com/neovim/neovim/blob/ec7453804421f738955a5f4460f44172318d42ff/src/nvim/plines.c#L550) which would handle overlong '^A' as a double-width character, [`ascii_iswhite(*ptr)`](https://github.com/neovim/neovim/blob/ec7453804421f738955a5f4460f44172318d42ff/src/nvim/edit.c#L4353)). My version handles them correctly only in some places.
3. Potential bug in [`coladvance2()`](https://github.com/neovim/neovim/blob/ec7453804421f738955a5f4460f44172318d42ff/src/nvim/cursor.c#L147): the `head` variable should probably be reset on every iteration of the loop, like in [`win_line()`](https://github.com/neovim/neovim/blob/ec7453804421f738955a5f4460f44172318d42ff/src/nvim/drawline.c#L1355). `win_lbr_chartabsize()` might not reset `*headp` if it calls `win_chartabsize()` or `win_nolbr_chartbasize()`, so `*headp` might contain 'head' value of one of the previous characters. This would be the case if wrapping is enabled and a double-width character intersect window edge. Then it would set `head` to 1 and it would stay 1 for all characters after it. 
4. 'win_lbr_chartbasize()` recomputed the same width for tab if was displayed like ^I and was moved by virtual text. I changed the condition to check for special tab behavior.

### Benchmarks

To verify that my changes actually improve performance, I wrote a Lua script that runs `screenpos()` 10'000 times in a loop and measures the difference in time returned by os.clock(). I don't know how GC and JIT affect the measurements, but most results are within 1% of the average for the test. Both the original and modified versions were compiled in release mode with Clang 14 and GCC 11.4 on Ubuntu x86-64 (WSL). Screen size is 224x57, wrapping is enabled.
There are 12 different tests in total: 4 with 10'000 characters in one line (long wrapped line), 4 with 100 lines with 220 characters each (long lines, none are wrapped), and 4 with 100 lines with 440 characters each (each line takes 2 screen rows). The results are measured for ascii 'i', and for 2 byte 'À', with and without `breakindent`. Time is in microseconds.
`screenpos()` was computed for the end of the last line, the file was scrolled to the end. 
I also did 4 additional test that measured the time it takes to insert 10'000 characters (with `vim.cmd"normal! 10000ix"`) for the same 2 characters with and without `breakindent`. Time is in **milliseconds**.

GCC. Results with `nobreakindent`:
||insert 10000 i|insert 10000 À|10'000 i|10'000 À|220 i| 220 À|440 i| 440 À|
|-|-|-|-|-|-|-|-|-|
|original|1292|6165|54|384|189|555|184|541|
|this|640|2385|41|160|58|204|55|198|

GCC. Results with `breakindent`:
||insert 10000 i|insert 10000 À|10'000 i|10'000 À|220 i| 220 À|440 i| 440 À|
|-|-|-|-|-|-|-|-|-|
|original|44673|99308|4314|10039|343|767|679|1149|
|this|4129|7604|276|559|341|608|344|600|

Clang, `nobreakindent`:
||insert 10000 i|insert 10000 À|10'000 i|10'000 À|220 i| 220 À|440 i| 440 À|
|-|-|-|-|-|-|-|-|-|
|original|1371|4145|58|264|203|453|195|373|
|this|630|2068|37|137|62|180|63|175|

Clang, `breakindent`:
||insert 10000 i|insert 10000 À|10'000 i|10'000 À|220 i| 220 À|440 i| 440 À|
|-|-|-|-|-|-|-|-|-|
|original|46048|94520|4324|10467|464|705|761|1181|
|this|4059|5666|264|383|330|448|360|494|