@VanaIgr 2023-12-31
I was looking at how to further optimize `win_lbr_chartabsize()` and it looks like some combinations of virtual text, double-width charaters and breakindent don't work as expected. For example, these 4 lines should probably end at the same virtual column:
![image](https://github.com/neovim/neovim/assets/65824523/9ef26c0a-179c-4920-8001-efd78604a8a3)
(white is regular text, red is virtual text. '?' is a double-width character. Virtual text in the third line contains the double-width char even though it is not displayed as red).

I assume that this is what's happening:
The first line's end column is different because ['>' character is counted](https://github.com/neovim/neovim/blob/6f6abd8c02f66a71f1ad46c4596cadd5709898f4/src/nvim/plines.c#L252) before applying breakindent, and the double-width character is not yet considered to be in the last column.
The third line's end column is different because the width of virtual text is [calculated without considering wrapping](https://github.com/neovim/neovim/blob/6f6abd8c02f66a71f1ad46c4596cadd5709898f4/src/nvim/api/extmark.c#L1160), and width of '>' for a double-width character in the last column is not added.