@zeertzjq 2023-12-31
> The first line's end column is different because ['>' character is counted](https://github.com/neovim/neovim/blob/6f6abd8c02f66a71f1ad46c4596cadd5709898f4/src/nvim/plines.c#L252) before applying breakindent

Are you referring to the 'breakindent' on the second screen line, or the third?