@VanaIgr 2023-12-31
First line in the file, second screen row.

Window width is 100. `breakindentopt=shift:2`, so it would add 2 cells to the next row of a line.

Virtual text will have a width of 197 cells. The first real character after virtual text is 2 cells. Total size is 199.

After computing total size, the function checks if `vcol + size - 2` is in last of its row. 0 + 199 - 2 is 197. Virtual column 197 is column 97 of a second row. Last column is 99, so head and size are not incremented.