@clason 2023-12-31
Impressive work! One small comment:

>  measures the difference in time returned by os.clock(). 

That function is not safe for (fast) benchmarking; prefer `vim.uv.hrtime()`, which has a higher resolution.