@VanaIgr 2024-01-01
> That function is not safe for (fast) benchmarking; prefer `vim.uv.hrtime()`, which has a higher resolution.

I had noticed that the scripts in 'test/benchmarks' used libuv's `hrtime()` instead of 'os.clock()' and ran some of my benchmarks with it to see if there was any significant difference, but the results were largely the same (average with `hrtime()` had around 10 decimal digits while `os.clock()` only had 1-4, but the results were within 1us of each other). 

Since I ran my benchmarks manually and couldn't find a way to copy anything from within neovim to the Windows clipboard, I used `os.clock()` to save some time on typing the results.

> And maybe consider adding these tests to `test/benchmark`, if possible?

I will write proper benchmarks and make them the new first commit.