@VanaIgr 2024-01-02
Do the optional s390x checks test if Neovim works on a big endian architecture?
If yes, than they are failing probably because `utf_ptr2CharInfo()` assumes the system is little-endian.