@dundargoc 2024-01-02
> Do the optional s390x checks test if Neovim works on a big endian architecture?

Yes.