@clason 2024-01-02
> If yes, than they are failing probably because utf_ptr2CharInfo() assumes the system is little-endian.

And yes. All failures are indeed relevant.