@bfredl 2024-01-05
> Is there a simpler way to execute Lua code inside Neovim and return the results to the benchmarks?

Yes, use `exec_lua` from helpers module.