@VanaIgr 2024-01-08
I also added `plines.h` to clint.py `check_includes_ignore` list, since it uses functions from `mbyte.h` to define `init_chartabsize()` and `adv_chartabsize()`. 
Since both are convenience functions, I can also inline them and remove the dependency on `mbyte.h`.