@VanaIgr 2024-01-13
I removed both functions. Now `plines.h` only includes mbyte_defs.