@VanaIgr 2024-01-13
I also found a couple of issues:

1. Blockwise put doesn't add correct number of spaces for multibyte characters.

The loop that determines the number of spaces to add to the pasted text ([link](https://github.com/neovim/neovim/blob/786089013d33666a2acb5ba86fb34c716a9cab26/src/nvim/ops.c#L3293)) advances the pointer by a single byte. This is not correct for multibyte characters, since all trailing bytes would be counted separately as 4 cell wide '\<xx\>'. This issue is also present in vim.
For example, copying the text at the top (blockwise, without virtualedit) and pasting it before the b's at the bottom while in blockwise visual mode would result in this:
![image](https://github.com/neovim/neovim/assets/65824523/2da78925-6bfc-4a1a-90c1-00a516d3aaa6) ![image](https://github.com/neovim/neovim/assets/65824523/d698bc2a-3bc8-4f60-aa77-6d7f0d22c3de)
This can be fixed by using `MB_PTR_ADV(cts.cts_ptr)` instead of `cts.cts_ptr++`, and changing the loop to iterate until the end of the string.

This still wouldn't work correctly for tabs, since all widths are computed at virtual column 0, but tab width changes depending on its position. I thought there also may be some issues if the tab's width is greater than window width and breakindent in enabled (since the size of breakindent in this case would be added to the total width of the pasted line) but for some reason this doesn't happen.

2. Lisp indenting doesn't use correct line number.

In [2 places](https://github.com/neovim/neovim/blob/786089013d33666a2acb5ba86fb34c716a9cab26/src/nvim/indent.c#L1275) in `get_lisp_indent()` , the 'lnum' argument for `init_chartabsize_arg()` is `(colnr_T)(that - line)` (which computes the index of the current byte, not the line number).
This can affect the results if the line at that position contains virtual text:
![image](https://github.com/neovim/neovim/assets/65824523/c819000f-3696-483c-a12a-7ca7e699e8bb)
![image](https://github.com/neovim/neovim/assets/65824523/f179859e-eb29-46f4-aa22-92dac898ecdb) ![image](https://github.com/neovim/neovim/assets/65824523/fbfef993-7ec7-40b3-81d7-45c0b0d4a5df)

There is also [one place](https://github.com/neovim/neovim/blob/786089013d33666a2acb5ba86fb34c716a9cab26/src/nvim/indent.c#L1270) in this function in which character is assumed to be 1 byte long and 1 cell wide. I don't know if this can cause any issues.