@justinmk 2024-01-13
Can you mention the user-facing consequences in `:help news`: https://github.com/neovim/neovim/blob/2dc439c672facaeb8e51ef6aa20efc0e7092eee2/runtime/doc/news.txt#L129 (e.g. "performance of 'breakindent', foo, bar, etc., is now 3x faster for long lines, etc, etc"

Also would be ideal if the commit message contains the explanation from the PR description, especially the list of problems addressed.