@clason 2024-01-16
Yes, fixing the endian issue that makes the test fail (see @bfredl's comment above).