@bfredl 2024-01-16
Hmm, there might be no more s3900 failures than those known to exist on master, which is also 8+3, see https://github.com/neovim/neovim/pull/26950