@clason 2024-01-16
Oh, right, I should have looked at the details first. Yes, the original (relevant) s390x failures seem to be gone. 🚀